Hacky fix for clock
We discovered that despite the Ethernet intialization functions properly
setting the SPI clock rate of the Ethernet to 8Mhz, the Lora initialization
functions does *something* to the system clocks that happens to make the
Ethernet SPI run at 24 Mhz...

Unfortunately, the ENC28J60 is remarkably *almost* able to work properly at
this speed, despite it being 3x the max, but the ways it doesn't work tend
to be nonsensical unexplainable errors.

We fixed this after several hours, then proceeded to *un*-fix it by forgetting
to commit resulting in us rediscovering the bug after another several hours
trying to debug another subtle issue.

(The first obvious one being that received Frames were wrong, the more subtle
second being that the output interrupt line was not properly resetting after
receiving)