Set up rudimentary system for interrupt-based handling!
Note: This currently has a very nasty potential race condition in it due
to our lack of atomic operations:
In check_set_eth_pending, if the interrupt line is high (inactive) when we
call the GPIO function, then the function is pre-empted by the ISR for the Eth
before we actually set the value, we will set eth_available to true before
immediately changing it to false as we return into check_set_eth_pending.

If this were to hypothetically happen, we will block for forever without
making progress.

This should probably be fixed at some point :)