Correctly enable promiscuous mode on ENC28J60 to receive all traffic
The subsequent calls to enable various receiving on the Ethernet had the side
effect of *disabling* promiscuous mode.

As a result, we dropped all unicast traffic instead of forwarding it along,
causing very obvious problems in being able to send traffic end-to-end :)

Just deleting the subsequent calls amusingly leaves us where we wanted to be,
with us receiving absolutely everything.